Refactoring attention, moe and ffn blocks
